build 170507. vs2015 compliance

+++ uilayer.h
@@ -76,7 +76,7 @@ private:
 
 
 // members:
-    GLuint m_fontbase{ -1 }; // numer DL dla znaków w napisach
+    GLuint m_fontbase{ (GLuint)-1 }; // numer DL dla znaków w napisach
     float m_progress{ 0.0f };                // percentage of filled progres bar, to indicate lengthy operations.
     float m_subtaskprogress{ 0.0f };                // percentage of filled progres bar, to indicate lengthy operations.
     texture_manager::size_type m_background; // path to texture used as the background. size depends on mAspect.
